refactor(views): migrate playlists to per-folder layout (commit 3/7)
Split src/views/playlists.rs (1318 lines) into src/views/playlists/{mod,update,view}.rs
following the genres template:

- mod.rs: types (Page, Message, Action, ViewData, PlaylistContextEntry),
  trait impls (HasCommonAction via macro with no_center, ViewPage explicit),
  Default + new()
- update.rs: impl PlaylistsPage { fn update }
- view.rs: impl PlaylistsPage { fn view, fn render_playlist_row,
  fn render_track_row } + the column-visibility helpers
  (playlists_song_count_visible / playlists_duration_visible /
  playlists_updated_at_visible) and the playlist context-menu helpers
  (playlist_context_entries / playlist_entry_view)

Macro consumers:
- define_view_columns! for PlaylistsColumn / Visibility (6 columns).
- impl_has_common_action!(PlaylistsAction, no_center { NavigateAndExpandArtist })
  — first use of the no_center variant since playlists has no
  CenterOnPlaying message (no "currently playing" concept on the playlist
  row level).

PlaylistContextEntry stays in mod.rs because it's referenced from all
three files (Message variant, update match, view rendering).

No behavioral change. Public API surface unchanged. cargo build / clippy
--all-targets -D warnings / nightly fmt --check / test --workspace all
clean (662 + 368 + 1 doctest).
